Move ALLOWED_SORTS and SIMPLY_SORTED to constants
